@@ -68,15 +68,8 @@
     "@angular/language-service": "16.0.4",
     "@commitlint/cli": "12.1.4",
     "@commitlint/config-conventional": "12.1.4",
-    "@jscutlery/semver": "2.26.0",
+    "@jscutlery/semver": "2.30.1",
     "@ngneat/spectator": "15.0.0",
-    "@nrwl/cli": "15.9.2",
-    "@nrwl/cypress": "15.9.2",
-    "@nrwl/eslint-plugin-nx": "15.9.2",
-    "@nrwl/jest": "15.9.2",
-    "@nrwl/js": "15.9.2",
-    "@nrwl/linter": "15.9.2",
-    "@nrwl/workspace": "15.9.2",
     "@nx/angular": "16.3.2",
     "@types/command-line-args": "5.2.0",
     "@types/find-duplicated-property-keys": "1.1.0",
@@ -85,9 +78,9 @@
     "@types/jasmine": "4.0.3",
     "@types/jest": "29.5.0",
     "@types/node": "16.7.2",
-    "@typescript-eslint/eslint-plugin": "5.57.0",
-    "@typescript-eslint/parser": "5.57.0",
-    "cypress": "8.7.0",
+    "@typescript-eslint/eslint-plugin": "5.59.11",
+    "@typescript-eslint/parser": "5.59.11",
+    "cypress": "12.14.0",
     "dotenv": "10.0.0",
     "eslint": "8.15.0",
     "eslint-config-prettier": "8.3.0",
@@ -107,7 +100,7 @@
     "karma-jasmine-html-reporter": "2.0.0",
     "lint-staged": "11.1.2",
     "ng-packagr": "16.0.1",
-    "nx": "15.9.2",
+    "nx": "16.3.2",
     "postcss": "8.4.21",
     "postcss-import": "14.1.0",
     "postcss-preset-env": "8.2.0",
@@ -116,7 +109,13 @@
     "ts-jest": "29.1.0",
     "ts-mocks": "3.0.1",
     "ts-node": "10.9.1",
-    "typescript": "~4.9.5"
+    "typescript": "5.0.4",
+    "@nx/workspace": "16.3.2",
+    "@nx/eslint-plugin": "16.3.2",
+    "@nx/cypress": "16.3.2",
+    "@nx/linter": "16.3.2",
+    "@nx/js": "16.3.2",
+    "@nx/jest": "16.3.2"
   },
   "config": {
     "commitizen": {
