@@ -11,8 +11,8 @@
 // This function is called when a project is opened or re-opened (e.g. due to
 // the project's config changing)
 
-module.exports = (on, config) => {
+export default function() {
   // `on` is used to hook into various events Cypress emits
   // `config` is the resolved Cypress config
   // Preprocess Typescript file using Nx helper
-};
+}
